add dia

+++ lib/models/pulse/user_income_total.dart
@@ -6,8 +6,8 @@ part 'user_income_total.g.dart';
 @freezed
 class UserIncomeTotal with _$UserIncomeTotal {
   factory UserIncomeTotal({
-    @JsonKey(name: 'today') @Default(0) int today,
-    @JsonKey(name: 'total') @Default(0) int total,
+    @JsonKey(name: 'today') @Default(0) double today,
+    @JsonKey(name: 'total') @Default(0) double total,
   }) = _UserIncomeTotal;
 
   factory UserIncomeTotal.fromJson(Map<String, dynamic> json) =>
